Consolidate virtual machine code under compliance/virtualmachines/relay
Moves relay-related code from compliance/virtualmachine to
compliance/virtualmachines/relay to eliminate the duplicate
directory structure and better organize the codebase.
