+++ compliance/virtualmachines/relay/relay_test.go
@@ -1,4 +1,4 @@
-package virtualmachine
+package relay
 
 import (
 	"context"
